+++ test/verify_dry_run.test.sh
@@ -47,7 +47,8 @@ printf '%s\n' '<script src="/pve2/js/pvemanagerlib.js"></script>' '</head>' '</b
 printf '%s\n' "Ext.define('PVE.form.ViewSelector');" "Ext.define('PVE.tree.ResourceTree');" \
     "Ext.define('PVE.node.StatusView');" "Ext.define('PVE.panel.Config');" \
     "Ext.define('PVE.sdn.VnetEdit');" "Ext.define('PVE.sdn.SubnetView');" \
-    "Ext.define('PVE.sdn.VnetACLView');" > "$PVE_MANAGER_JS"
+    "Ext.define('PVE.sdn.VnetACLView');" "Ext.define('PVE.dc.CmdMenu');" \
+    "Ext.define('PVE.node.CmdMenu');" > "$PVE_MANAGER_JS"
 printf '%s\n' '        my $dinfo = df('\''/'\'', 1);' > "$NODES_PM"
 printf '%s\n' 'my $observed = {' '};' > "$PVE_CLUSTER_PM"
 printf '%s\n' 'package PVE::API2;' 'use base qw(PVE::RESTHandler);' '1;' > "$PVE_API2_PM"